+++ lib/GAME_H/element.hpp
@@ -22,12 +22,12 @@ public:
 
     //getter & setter pos
     glm::vec3 getPos() const {return m_pos;};
-    float getPosX() const {return m_pos.x;};
-    float getPosY() const {return m_pos.y;};
-    float getPosZ() const {return m_pos.z;};
-    void setPosX(const float posX) {m_pos.x = posX;};
-    void setPosY(const float posY) {m_pos.y = posY;};
-    void setPosZ(const float posZ) {m_pos.z = posZ;};
+    int getPosX() const {return m_pos.x;};
+    int getPosY() const {return m_pos.y;};
+    int getPosZ() const {return m_pos.z;};
+    void setPosX(const int posX) {m_pos.x = posX;};
+    void setPosY(const int posY) {m_pos.y = posY;};
+    void setPosZ(const int posZ) {m_pos.z = posZ;};
 
     //virtual fonction de collision
     bool possibleAdd; //if we can add another element (obsacle, money)
